Release 0.4.1
Release 0.4.1
- Added the Ponderosa, Wild Blue Industries’ top of the line
colonization module. The Ponderosa can be stored in a KAS container (it
takes up 90 KAS units of space) and stored on a standard KAS container
bay. To deploy the Ponderosa, simply grab the module while on EVA, walk
or fly over to the desired site, drop the module, and deploy it.
Optionally, you can attach it to the ground before deployment, but
you’ll need to re-orient the module first.
- Added the Portable Crew Tube (PCT). The PCT is just like the
Flex-O-Tube except that the PCT has a more stock-like appearance and it
can be stored on the MBU for easy attachment.
- Repositioned the location where new spacecraft are spawned by the
Skydock250 and Skydock375. Hopefully that’ll reduce the chances of the
docks causing an explosion when spawning new craft.
- Added MM config file to hide unused MKS/OKS parts. They’re still
around, just not shown in the editor.
- Added MM config file for KSOS modules: The cupola can now become a
Colony Command Center; the station power module can now become a PDU;
the KerbaLab can now become a Multipurpose Colony Module; the
hydroponics module may be used as an efficiency part for the Aeroponics
module; and the habitat module may be used as an efficiency part for the
Kerbitat.
NOTE: You might need to delete your ModuleManager.CONFIGCACHE and
ModuleManager.ConfigSHA.CONFIGSHA file, located in the GameData folder,
for the MM config changes to take effect.
- Added OSE workshop to the Fabricator template. Now you can print KAS
grabbable and storable parts! OSE Workshop by ObiVanDamme downloaded
separately. Link:
http://forum.kerbalspaceprogram.com/threads/108234-0-90-OSE-Workshop-KAS-Addon-%28v0-2-0-2015-01-26%29
NOTE: If you don't have the OSE Workshop, no worries, Multipurpose
Colony Modules won't complain (but you are missing out...)